// src/constants/enums.ts
export enum Routes {
  Root = "/",
  Auth = "auth",
  Admin = "admin",
}

export enum Pages {
  Signin = "signin",
  Blogs = "blogs",
  Projects = "projects",
  New = "new",
  Edit = "edit",
}

export enum Environments {
  Prod = "production",
  Dev = "development",
}

export enum InputTypes {
  TEXT = "text",
  EMAIL = "email",
  PASSWORD = "password",
  NUMBER = "number",
  DATE = "date",
  TIME = "time",
  DATE_TIME_LOCAL = "datetime-local",
  CHECKBOX = "checkbox",
  RADIO = "radio",
  SELECT = "select",
  TEXTAREA = "textarea",
  FILE = "file",
  IMAGE = "image",
  COLOR = "color",
  RANGE = "range",
  TEL = "tel",
  URL = "url",
  SEARCH = "search",
  MONTH = "month",
  WEEK = "week",
  HIDDEN = "hidden",
  MULTI_SELECT = "multi select",
}